float endian functions moved to separate file
