+++ src/Makefile
@@ -23,11 +23,13 @@ testsuite.o : tst/testsuite.cpp $(GTEST_HEADERS)
 	@ echo "Compiling Tests..."
 	@ g++ $(CPPFLAGS) -isystem $(GTEST_DIR)/include $^ -c
 
-testsuite : testsuite.o gtest-all.o gtest_main.o hydrologyFunctions.o hydrologyParameters.o hydrology.o
+testsuite : testsuite.o gtest-all.o gtest_main.o hydrologyFunctions.o hydrologyParameters.o hydrology.o floatEndian.o
 	g++ $(CPPFLAGS) -lpthread -lopencv_core -lopencv_imgproc -lopencv_highgui $^ -o $@
 
 test : testsuite
 	@ ./testsuite
+
+floatEndian.o : floatEndian.cpp
 
 hydrology.o : hydrology.cpp hydrology.hpp point.hpp
 
@@ -37,8 +39,8 @@ hydrologyParameters.o : hydrologyParameters.cpp raster.hpp hydrology.hpp
 
 hydrologyFunctions.o : hydrologyFunctions.cpp hydrologyFunctions.hpp hydrology.hpp hydrologyParameters.hpp
 
-buildRivers: hydrologyFunctions.o hydrologyParameters.o buildRivers.o hydrology.o
-	g++ hydrologyFunctions.o buildRivers.o hydrologyParameters.o hydrology.o $(CPPFLAGS) -lopencv_core -lopencv_imgproc -lopencv_highgui -o $@
+buildRivers: hydrologyFunctions.o hydrologyParameters.o buildRivers.o hydrology.o floatEndian.o
+	g++ $^ $(CPPFLAGS) -lopencv_core -lopencv_imgproc -lopencv_highgui -o $@
 
 clean :
 	@ rm -vf buildRivers testsuite *.gch *.o
