+++ src/floatEndian.hpp
@@ -1,0 +1,8 @@
+#ifndef FLOAT_ENDIAN_H
+#define FLOAT_ENDIAN_H
+
+float float_swap(float value);
+
+float float_tobe(float value);
+
+#endif
